SafetyManagerApp/requirements.txt
django-webpack-loader==0.7.0 
asgiref==3.2.7  
bson==0.5.8
certifi==2020.4.5.1
dataclasses==0.6
Django==2.0
django-mongodb-engine==0.6.0
djangotoolbox==1.8.0
djongo==1.2.30
mongoengine==0.20.0
Pillow==7.1.2
pip==20.1
pymongo==3.2
python-dateutil==2.8.1
pytz==2018.5
setuptools==46.1.3
six==1.14.0
sqlparse==0.2.3
wheel==0.34.2
wincertstore==0.2 
djangorestframework == 3.11.0